[SDK/INCLUDE] Introduce boot display discovery helper functions
Their purpose is to simplify the way a compatible boot video driver
and a generic framebuffer miniport driver can retrieve this information.

In this iteration, the data is retrieved only from the kernel loader-block
ARC tree, via the `KeFind(Next)ConfigurationEntry()` routines.
will be used by the boot video driver (since it runs early in the boot process).

Other methods, suitable for retrieving this information later in the
boot process, will be added in the future.